Create Arduino INI file
Struggling with what to put on the lcd.print's to get the LCD to display the target voltage first and then the plasma incoming voltage below that on the lcd under the writing ("Voltage Reading" ) I will eventually add a 10K pot to adjust voltage on A0, just don't have one today. Its in the mail.